+++ NEWS.md
@@ -10,6 +10,8 @@ repositories mentioned in `README.md`.
 v0.x.x - 2025-xx-xx
 -------------------
 
+- Add the `bench` command, executes various benchmarks and reports the 
+  results.
 - Remove errno messages if invalid numbers were specified certain 
   places.
 - Remove errno message on FreeBSD when `--count` and `--seed` got empty